major path refactor, submodules messed up
