+++ data/deployments/Kampala/manifest.json
