+++ data/forms/buildings_with_pois.xml
