+++ jekyll/README.md
@@ -1,0 +1,9 @@
+# Hi @rrlara !!!!
+
+This is where the Jekyll source files will live. We will publish the output to `/pages` instead of `_site`.
+
+Right now we have some basic sites in the `/pages` directory, and we're going to want to integrate what
+we've got into the Jekyll setup. _Discussion needed..._
+
+PS: We're going to want to move `/docs` into `/jekyll` so that they get published as well. Let's figure that out
+once we've got the boilerplate Jekyll site set up.
