+++ settings.js
@@ -1,7 +1,7 @@
 module.exports = {
     name: 'OpenMapKit Server',
     description: 'OpenMapKit Server is the lightweight server component of OpenMapKit that handles the collection and aggregation of OpenStreetMap and OpenDataKit data.',
-
     port: 3210,
-    publicDir: __dirname + '/public'
+    dataDir: __dirname + '/data',
+    pagesDir: __dirname + '/pages'
 };
