+++ test/osm/filter-osm.test.js
@@ -1,7 +1,7 @@
 var test = require('tape');
 var fs = require('fs');
 var libxml = require('libxmljs');
-var filterOsm = require('../../osm/filter-osm');
+var filterOsm = require('.././filter-osm');
 
 /**
  * Get together an array of the simple set of OSM files to test.
